+++ .idea/dataSources.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="DataSourceManagerImpl" format="xml" multifile-model="true">
-    <data-source source="LOCAL" name="employesDataSource" uuid="5c01fdb8-3e3a-4798-b83e-cc65db244df2">
+    <data-source source="LOCAL" name="projectDataSource" uuid="d16d3a70-8523-4d12-969c-28a83a249970">
       <driver-ref>mysql.8</driver-ref>
       <synchronize>true</synchronize>
       <jdbc-driver>com.mysql.cj.jdbc.Driver</jdbc-driver>
-      <jdbc-url>jdbc:mysql://localhost:3306/employes</jdbc-url>
+      <jdbc-url>jdbc:mysql://localhost:3306/apprentrack</jdbc-url>
       <working-dir>$ProjectFileDir$</working-dir>
     </data-source>
   </component>
