--- .idea/sqldialects.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="SqlDialectMappings">
-    <file url="file://$PROJECT_DIR$/src/main/java/fr/efrei/apptrack/utils/Employes - Script SQL - MySQL.sql" dialect="MySQL" />
-  </component>
-</project>
