--- src/main/java/fr/efrei/apptrack/utils/Employes - Script SQL - MySQL.sql
@@ -1,30 +1,0 @@
-/**** CIBLE : MySQL ****/
-
-/*On supprime les tables si elles existent */
-DROP TABLE IF EXISTS EMPLOYES;
-
-CREATE TABLE EMPLOYES(
-	ID INT NOT NULL AUTO_INCREMENT,
-	NOM VARCHAR(25),
-	PRENOM VARCHAR(25),
-    TELDOM VARCHAR(10),
-	TELPORT VARCHAR(10),
-	TELPRO VARCHAR(10),
-	ADRESSE VARCHAR(150),
-	CODEPOSTAL VARCHAR(15),
-	VILLE VARCHAR(25),
-	EMAIL VARCHAR(25),
-    PRIMARY KEY (ID)
-);
-
-/*Insertion de 4 membres*/
-INSERT INTO EMPLOYES(NOM,PRENOM,TELDOM,TELPORT,TELPRO,ADRESSE,CODEPOSTAL,VILLE,EMAIL) VALUES
-('Bond','James','0123456789','0612345678','0698765432','2 avenue 007','92700','Colombes','jbond@gmail.com');
-INSERT INTO EMPLOYES(NOM,PRENOM,TELDOM,TELPORT,TELPRO,ADRESSE,CODEPOSTAL,VILLE,EMAIL) VALUES
-('Jones','Indiana','0145362787','0645362718','0611563477','10 rue des Aventuriers','92270','Bois-colombes','ijones@gmail.com');
-INSERT INTO EMPLOYES(NOM,PRENOM,TELDOM,TELPORT,TELPRO,ADRESSE,CODEPOSTAL,VILLE,EMAIL) VALUES
-('Bourne','Jason','0187665987','0623334256','0654778654','65 rue Agent Secret Perdu','92700','Colombes','jbourne@yahoo.fr');
-INSERT INTO EMPLOYES(NOM,PRENOM,TELDOM,TELPORT,TELPRO,ADRESSE,CODEPOSTAL,VILLE,EMAIL) VALUES
-('Stark','Arya','0187611987','0783334256','0658878654','6 rue Sans-Nom','75016','Paris','astark@nord.com');
-INSERT INTO EMPLOYES(NOM,PRENOM,TELDOM,TELPORT,TELPRO,ADRESSE,CODEPOSTAL,VILLE,EMAIL) VALUES
-('Lanister','Cersei','0187384987','0622494256','0674178654','5 bvd des Reines','21000','Dijon','clanister@mail.co.uk');
